+++ common/post-fs-data.sh
@@ -5,17 +5,21 @@
 
 
 main() {
-  # verbose engine
-  logsDir=/data/media/adk/logs
-  newLog=$logsDir/main_verbose_log.txt
-  oldLog=$logsDir/main_verbose_previous_log.txt
-  [[ -d $logsDir ]] || mkdir -p -m 777 $logsDir
+  modID=adk
+  modPath=${0%/*}
+  functionName=main
+  modData=/data/media/$modID
+
+  # verbosity engine
+  logsDir=$modData/logs
+  newLog=$logsDir/${functionName}_verbose_log.txt
+  oldLog=$logsDir/${functionName}_verbose_previous_log.txt
+  [[ -d $logsDir ]] || mkdir -p $logsDir
   [[ -f $newLog ]] && mv $newLog $oldLog
   set -x 2>>$newLog
 
-  modPath=${0%/*}
   . $modPath/adk.sh
-  main
+  $functionName
 }
 
 (main) &