+++ Changelog.md
@@ -1,2 +1,7 @@
+**2018.7.29 (201807290)**
+- Auto-detect whether adk should go to bin or xbin dir to avoid bootloops
+- General optimizations
+- Updated documentation
+
 **2018.7.28**
 - Initial release